+++ python/GafferTest/ComputeNodeTest.py
@@ -740,5 +740,14 @@ class ComputeNodeTest( GafferTest.TestCase ) :
 		self.assertNotEqual( h2, h1 )
 		self.assertEqual( n.plug.getValue(), "out2" )
 
+	def testInterleavedEditsAndComputes( self ) :
+
+		n = GafferTest.AddNode()
+
+		with Gaffer.DirtyPropagationScope() :
+			for i in range( 0, 100 ) :
+				n["op1"].setValue( i )
+				self.assertEqual( n["sum"].getValue(), i )
+
 if __name__ == "__main__":
 	unittest.main()